fix static deploy
