+++ sofa-ark-parent/core-impl/container/src/main/java/com/alipay/sofa/ark/container/pipeline/DeployBizStage.java
@@ -51,8 +51,7 @@ public class DeployBizStage implements PipelineStage {
         eventAdminService.sendEvent(new AfterFinishDeployEvent());
     }
 
-    public void processStaticBiz(PipelineContext pipelineContext)
-                                                                                throws ArkRuntimeException {
+    public void processStaticBiz(PipelineContext pipelineContext) throws ArkRuntimeException {
         String[] args = pipelineContext.getLaunchCommand().getLaunchArgs();
         bizDeployService.deploy(args);
     }
