+++ sofa-ark-parent/support/ark-springboot-integration/ark-springboot-starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,6 @@
 org.springframework.context.ApplicationListener=\
 com.alipay.sofa.ark.springboot.listener.ArkApplicationStartListener,\
-com.alipay.sofa.ark.springboot.listener.ArkDeployBizListener,\
+com.alipay.sofa.ark.springboot.listener.ArkDeployStaticBizListener,\
 com.alipay.sofa.ark.springboot.listener.PropertiesResetListener
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.alipay.sofa.ark.springboot.ArkAutoConfiguration
